+++ py/bitbox02/bitbox02/bitbox02/bitbox02.py
@@ -561,9 +561,37 @@ class BitBox02(BitBoxCommonAPI):
         request.sign_message.CopyFrom(
             btc.BTCSignMessageRequest(coin=coin, script_config=script_config, msg=msg)
         )
-        sig = self._btc_msg_query(request, expected_response="sign_message").sign_message.signature
 
-        sig, recid = sig[:64], sig[64]
+        supports_antiklepto = self.version >= semver.VersionInfo(9, 5, 0)
+        if supports_antiklepto:
+            host_nonce = os.urandom(32)
+
+            request.sign_message.host_nonce_commitment.commitment = antiklepto_host_commit(
+                host_nonce
+            )
+            signer_commitment = self._btc_msg_query(
+                request, expected_response="antiklepto_signer_commitment"
+            ).antiklepto_signer_commitment.commitment
+
+            request = btc.BTCRequest()
+            request.antiklepto_signature.CopyFrom(
+                antiklepto.AntiKleptoSignatureRequest(host_nonce=host_nonce)
+            )
+
+            signature = self._btc_msg_query(
+                request, expected_response="sign_message"
+            ).sign_message.signature
+            antiklepto_verify(host_nonce, signer_commitment, signature[:64])
+
+            if self.debug:
+                print(f"Antiklepto nonce verification PASSED")
+
+        else:
+            signature = self._btc_msg_query(
+                request, expected_response="sign_message"
+            ).sign_message.signature
+
+        sig, recid = signature[:64], signature[64]
 
         # See https://github.com/spesmilo/electrum/blob/84dc181b6e7bb20e88ef6b98fb8925c5f645a765/electrum/ecc.py#L521-L523
         compressed = 4  # BitBox02 uses only compressed pubkeys
@@ -704,6 +732,29 @@ class BitBox02(BitBoxCommonAPI):
                 data=data,
             )
         )
+
+        supports_antiklepto = self.version >= semver.VersionInfo(9, 5, 0)
+        if supports_antiklepto:
+            host_nonce = os.urandom(32)
+
+            request.sign.host_nonce_commitment.commitment = antiklepto_host_commit(host_nonce)
+            signer_commitment = self._eth_msg_query(
+                request, expected_response="antiklepto_signer_commitment"
+            ).antiklepto_signer_commitment.commitment
+
+            request = eth.ETHRequest()
+            request.antiklepto_signature.CopyFrom(
+                antiklepto.AntiKleptoSignatureRequest(host_nonce=host_nonce)
+            )
+
+            signature = self._eth_msg_query(request, expected_response="sign").sign.signature
+            antiklepto_verify(host_nonce, signer_commitment, signature[:64])
+
+            if self.debug:
+                print(f"Antiklepto nonce verification PASSED")
+
+            return signature
+
         return self._eth_msg_query(request, expected_response="sign").sign.signature
 
     def eth_sign_msg(
@@ -716,6 +767,29 @@ class BitBox02(BitBoxCommonAPI):
         request = eth.ETHRequest()
         # pylint: disable=no-member
         request.sign_msg.CopyFrom(eth.ETHSignMessageRequest(coin=coin, keypath=keypath, msg=msg))
+
+        supports_antiklepto = self.version >= semver.VersionInfo(9, 5, 0)
+        if supports_antiklepto:
+            host_nonce = os.urandom(32)
+
+            request.sign_msg.host_nonce_commitment.commitment = antiklepto_host_commit(host_nonce)
+            signer_commitment = self._eth_msg_query(
+                request, expected_response="antiklepto_signer_commitment"
+            ).antiklepto_signer_commitment.commitment
+
+            request = eth.ETHRequest()
+            request.antiklepto_signature.CopyFrom(
+                antiklepto.AntiKleptoSignatureRequest(host_nonce=host_nonce)
+            )
+
+            signature = self._eth_msg_query(request, expected_response="sign").sign.signature
+            antiklepto_verify(host_nonce, signer_commitment, signature[:64])
+
+            if self.debug:
+                print(f"Antiklepto nonce verification PASSED")
+
+            return signature
+
         return self._eth_msg_query(request, expected_response="sign").sign.signature
 
     def reset(self) -> bool: