+++ src/rust/bitbox02-rust/src/hww/api/ethereum.rs
@@ -28,6 +28,44 @@ use super::Error;
 use pb::eth_request::Request;
 use pb::eth_response::Response;
 
+use core::convert::TryInto;
+
+/// Like `hww::next_request`, but for Ethereum requests/responses.
+pub async fn next_request(response: Response) -> Result<Request, Error> {
+    let request = crate::hww::next_request(pb::response::Response::Eth(pb::EthResponse {
+        response: Some(response),
+    }))
+    .await?;
+    match request {
+        pb::request::Request::Eth(pb::EthRequest {
+            request: Some(request),
+        }) => Ok(request),
+        _ => Err(Error::InvalidState),
+    }
+}
+
+/// Sends the `signer_nonce_commitment` to the host and waits for the next request, which has to be a
+/// `AntiKleptoSignatureRequest` message containing the host nonce.
+pub async fn antiklepto_get_host_nonce(
+    signer_nonce_commitment: [u8; 33],
+) -> Result<[u8; 32], Error> {
+    let request = next_request(Response::AntikleptoSignerCommitment(
+        pb::AntiKleptoSignerCommitment {
+            commitment: signer_nonce_commitment.to_vec(),
+        },
+    ))
+    .await?;
+    match request {
+        Request::AntikleptoSignature(pb::AntiKleptoSignatureRequest { host_nonce }) => {
+            Ok(host_nonce
+                .as_slice()
+                .try_into()
+                .or(Err(Error::InvalidInput))?)
+        }
+        _ => Err(Error::InvalidState),
+    }
+}
+
 /// Handle a Ethereum protobuf api call.
 ///
 /// Returns `None` if the call was not handled by Rust, in which case it should be handled by
@@ -37,5 +75,6 @@ pub async fn process_api(request: &Request) -> Option<Result<Response, Error>> {
         Request::Pub(ref request) => Some(pubrequest::process(request).await),
         Request::SignMsg(ref request) => Some(signmsg::process(request).await),
         Request::Sign(ref request) => Some(sign::process(request).await),
+        Request::AntikleptoSignature(_) => Some(Err(Error::InvalidInput)),
     }
 }